requirements.txt
Pillow==8.2.0
paramiko==2.7.2
python-dotenv==0.15.0